AllFarmers get all farmers api called
