+++ frontend/src/pages/AllFarmers.jsx
@@ -1,3 +1,7 @@
+import React, {useState, useEffect} from 'react';
+import axios from 'axios';
+import { Link } from 'react-router-dom';
+
 import { ChevronDownIcon } from '@heroicons/react/solid'
 
 const people = [
@@ -7,6 +11,16 @@ const people = [
 ]
 
 export default function AllFarmers() {
+
+    const [farmers, setFarmers] = useState([]);
+
+    useEffect(() => {
+        axios.get('http://localhost:8000/api/farmers/')
+            .then((res) => {
+                setFarmers(res.data);
+            });
+    }, []);
+
     return (
         <div className="px-4 sm:px-6 lg:px-8">
             <div className="sm:flex sm:items-center">
@@ -68,16 +82,16 @@ export default function AllFarmers() {
                                 </tr>
                                 </thead>
                                 <tbody className="divide-y divide-gray-200 bg-white">
-                                {people.map((person) => (
-                                    <tr key={person.email}>
+                                {farmers.map((farmer) => (
+                                    <tr key={farmer.email}>
                                         <td className="whitespace-nowrap py-4 pl-4 pr-3 text-sm font-medium text-gray-900 sm:pl-6">
-                                            {person.name}
+                                            {farmer.name}
                                         </td>
-                                        <td className="whitespace-nowrap px-3 py-4 text-sm text-gray-500">{person.email}</td>
-                                        <td className="whitespace-nowrap px-3 py-4 text-sm text-gray-500">{person.role}</td>
+                                        <td className="whitespace-nowrap px-3 py-4 text-sm text-gray-500">{farmer.email}</td>
+                                        <td className="whitespace-nowrap px-3 py-4 text-sm text-gray-500">{farmer.hectare}</td>
                                         <td className="relative whitespace-nowrap py-4 pl-3 pr-4 text-right text-sm font-medium sm:pr-6">
                                             <a href="#" className="text-indigo-600 hover:text-indigo-900">
-                                                Edit<span className="sr-only">, {person.name}</span>
+                                                Edit<span className="sr-only">, {farmer.name}</span>
                                             </a>
                                         </td>
                                     </tr>
